+++ src/cmp.stub.js
@@ -23,20 +23,20 @@ if (window.__cmp !== undefined) {
 const isEmpty = (obj) => {
 	return Object.keys(obj).length === 0 && obj.constructor === Object;
 };
-// const deleteCookie = (name, path, domain) => {
-// 	console.debug('NDM Delete: ', name, path, domain);
-// 	document.cookie = name + "=" +
-// 		((path) ? ";path="+path:"")+
-// 		((domain)?";domain="+domain:"") +
-// 		";expires=Thu, 01 Jan 1970 00:00:01 GMT";
-// };
+const deleteCookie = (name, path, domain) => {
+	console.debug('NDM Delete: ', name, path, domain);
+	document.cookie = name + "=" +
+		((path) ? ";path="+path:"")+
+		((domain)?";domain="+domain:"") +
+		";expires=Thu, 01 Jan 1970 00:00:01 GMT";
+};
 
 getTCData(window, (tcData) => {
 	// vvv temporary fix.
 	console.debug('NDM LegitimateInterests test', tcData);
 	if (tcData.gdprApplies && isEmpty(tcData.purpose.legitimateInterests)) {
 		console.warn('NDM Missing legitimateInterests', tcData);
-		// deleteCookie('euconsent-v2', '/', '.' + window.location.hostname);
+		deleteCookie('euconsent-v2', '/', '.' + window.location.hostname);
 	} else {
 		console.debug('NDM Find legitimateInterests or GDPR does not apply', tcData);
 	}